api/requirements.txt
google-generativeai==0.8.2
PyMuPDF==1.23.26
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0
elevenlabs==0.2.27
